+++ test/unit/helper/WebDriverIO_test.js
@@ -76,24 +76,6 @@ describe('WebDriverIO', function () {
           e.should.be.instanceOf(AssertionFailedError);
           e.inspect().should.include('web page');
       });
-  });
-
-  describe('see element : #seeElementInDOM, #dontSeeElementInDOM', () => {
-    it('should check elements are in the DOM', () => {
-      return wd.amOnPage('/form/field')
-        .then(() => wd.seeElementInDOM('input[name=name]'))
-        .then(() => wd.seeElementInDOM('//input[@id="name"]'))
-        .then(() => wd.dontSeeElementInDOM('#something-beyond'))
-        .then(() => wd.dontSeeElementInDOM('//input[@id="something-beyond"]'));
-    });
-  });
-
-  describe('see element : #seeElement, #dontSeeElement', () => {
-    it('should check elements are visible on the page', () => {
-      return wd.amOnPage('/form/field')
-        .then(() => wd.seeElementInDOM('input[name=email]'))
-        .then(() => wd.dontSeeElement('input[name=email]'))
-        .then(() => wd.dontSeeElement('#something-beyond'));
     });
   });
 
@@ -192,7 +174,7 @@ describe('WebDriverIO', function () {
     });
   });
 
-  describe.only('popup : #acceptPopup, #seeInPopup, #cancelPopup', () => {
+  describe('popup : #acceptPopup, #seeInPopup, #cancelPopup', () => {
     it('should accept popup window', () => {
       return wd.amOnPage('/form/popup')
         .then(() => wd.click('Confirm'))