+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 * [WebDriverIO] added `dragAndDrop` by @fabioel
 * [WebDriverIO] fixed `scrollTo` method by @sensone
 * [WebDriverIO] fixed `windowSize: maximize` option in config
+* [WebDriverIO] `seeElement` and `dontSeeElement` check element for visibility by @fabioel and @davertmik
+* [WebDriverIO] `seeElementInDOM`, `dontSeeElementInDOM` to check element exists on page.
 * fixed `within` block doesn't end in output not #79
+
 
 ## 0.2.8